requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Session==0.5.0
python-dotenv==1.0.0
google-generativeai==0.3.2
sendgrid==6.11.0
psycopg2-binary==2.9.10
requests==2.31.0
Werkzeug==3.0.1
gunicorn==21.2.0
razorpay==1.4.2
